Never reference rendr.server as a global
To enable this, we inject `dataAdapter` into the `apiProxy`
middleware, and also we attach `dataAdapter` to `req`, so that the
`syncer` can access. It would be nice to find an alternative to this, so
we don't bloat the `req`.

This is a necessary step towards enabling developers to run multiple
Rendr apps in the same Express app, which can be useful.