Better handling of error in Fetcher
Fetcher.fetchFromApi() provides error handler for Backbone's
Model.fetch(), and it expects a response body as a 2nd argument.
However, on the client-seid, it is an XHR object.

This patch fixes it, and adds a response body to an Error object,
because it may be helpful when Controllers decide a navigation.