Adding checks for routes using regexes.
the routes still need to be prepared a bit due to express using a leading `/` and backbone
not using the leading slash.  So even with this, the routes file has some logic:

```
//backbone style regex, no leading /
var routes =
[
  ['', 'getHome', 'home#index'],
  [/^foo.*/, 'getFoo', 'foo#index'],
  [/^bar.*/, 'getBar', 'bar#index']
];

module.exports = function(match) {
  routes.forEach( function(el, i){
    pattern = el[0];
    if(el[0] instanceof RegExp) {
      if(this.isServer) {
        //add the leading / for express if the regex is looking at the start of the path
        pattern = new RegExp(routeArray[0].toString().slice(1,-1).replace(/^\^/, '^\\\/'));
      }
    }
    match(pattern, el[1], el[2]);
  });
};
```

This logic could be moved into the `shared/base/router.js` buildRoutes:captureRoutes method:

```
  function captureRoutes() {
    var routeArray = _.toArray(arguments);

    if(routeArray[0] instanceof RegExp) {
      if(this.isServer) {
        //add the leading / for express if the regex is looking at the start of the path
        routeArray[0] = new RegExp(routeArray[0].toString().slice(1,-1).replace(/^\^/, '^\\\/'));
      }
    }
    routes.push(routeArray);
  }
```

but didn't want to suggest too much, just wanted to get regex routes to get though the flow.
Happy to update with the `captureRoutes` change if you see fit.