Adding ability to pass config/data to ViewEngine
The desire here is to be able to pass data into the __layout.hbs when it
is initially rendered on the server. It's data that we don't want to
go into the actual app as it's too big and only needs to be generated once when
the server starts up (eg: inline css).