Support 'spec.needsFetch === true' in Fetcher
As an easy way to force a specific resource to be fetched fresh, and not
from the ModelStore or CollectionStore.

    var spec = {
      model: {model: 'User', params: {id: 1234}, needsFetch: true}
    };
    this.app.fetch(spec, callback);

Previously, `spec.needsFetch` could be a function that accepts the model
data from the store and returns true or false.

Another way to force fetching is to use the `{readFromCache: false}`
option in `Fetcher#fetch`, but that option applies to the whole fetch,
not per-resource, like `spec.needsFetch` does.

    var spec = {
      model: {model: 'User', params: {id: 1234}}
    };
    this.app.fetch(spec, {readFromCache: false}, callback);