The try / catch is making me crazy
if the app_view in my app has an error, rendr quietly falls back to using it's own app_view.  Which is tricky to notice, as rendr's app_view does most of the work already.

But this is only a partial fix.  Getting the customAppView set on the app requires overriding the `initialize` method in the app's app_view, setting the `customAppView`, and then calling ` BaseApp.prototype.initialize.apply(this, arguments);` at the end of it.  I'm not thrilled about that as a solution, so I wanted to check in and get people's thoughts before adding a `preInitialize()` method, or if I'm missing a better way to get this into the base app.