Custom apiProxy server option
I have a situation where I don't need the apiProxy. Calls go directly
to an endpoint on the same server. This change seemed like the most
painless way of making that happen.

To get it working with no proxy and local api calls, I just need to do
the following.

1) Name the entries in dataAdapterConfig the version of the API "v1":
{...}

2) Override syncer.formatClientUrl in the Base Model/Collection to skip
adding the /-/ separator

3) Set config.apiProxy to noop

result is /api/v1/myresource -> app server = ok!

I could see the no-proxy being an option that does all of this.  It
prevents the middleware from hooking on and eliminates the addition of
the separator.