Make sure renderOnRefresh works after hydration
When the view is hydrated with data from the server, this.model/collection was
not yet set when ```initialize``` ran, so the view never started listening on
the model/collection's 'refresh' event.

Moving the event listener setup to ```parseOptions``` means the listener will
be set up every time the view is passed new options like model/collection.