make model/collection parsing an option in parseModelAndCollection
When parseModelAndCollection is called by the rendr-handlebars `view` method on the client-side there is not an app option present, so if the models parse method uses this.app it will error. Beyond that it seems unnecessary to call parse in this use-case which is just constructing the fetch summary for the placeholder div.

We do want the model/collection to be instantiated using parse when the view is attached, so pass `{ parse: true }` explicitly from `parseOptions`